requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
click==8.1.3
Django==4.0.4
django-bootstrap4==22.1
django-crispy-forms==1.14.0
django-ipware==4.0.2
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==9.1.1
soupsieve==2.3.2.post1
sqlparse==0.4.2
